+++ package/expo-package/src/handlers/index.ts
@@ -1,5 +1,1 @@
 export * from './compressImage';
-export * from './getLocalAssetUri';
-export * from './getPhotos';
-export * from './iOS14RefreshGallerySelection';
-export * from './oniOS14GalleryLibrarySelectionChange';