+++ package/native-package/src/index.js
@@ -2,18 +2,17 @@ import { Platform } from 'react-native';
 
 import { registerNativeHandlers } from 'stream-chat-react-native-core';
 
-import {
-  compressImage,
-  getLocalAssetUri,
-  getPhotos,
-  iOS14RefreshGallerySelection,
-  oniOS14GalleryLibrarySelectionChange,
-} from './handlers';
+import { compressImage } from './handlers';
 
 import {
   Audio,
   deleteFile,
+  getLocalAssetUri,
+  getPhotos,
+  iOS14RefreshGallerySelection,
+  oniOS14GalleryLibrarySelectionChange,
   pickDocument,
+  pickImage,
   saveFile,
   setClipboardString,
   shareImage,
@@ -32,6 +31,7 @@ registerNativeHandlers({
   iOS14RefreshGallerySelection,
   oniOS14GalleryLibrarySelectionChange,
   pickDocument,
+  pickImage,
   saveFile,
   SDK: 'stream-chat-react-native',
   setClipboardString,