+++ package/src/components/Attachment/Gallery.tsx
@@ -32,7 +32,7 @@ import {
 import { useTheme } from '../../contexts/themeContext/ThemeContext';
 import { useLoadingImage } from '../../hooks/useLoadingImage';
 import { isVideoPlayerAvailable } from '../../native';
-import type { DefaultStreamChatGenerics } from '../../types/types';
+import { DefaultStreamChatGenerics, FileTypes } from '../../types/types';
 import { getUrlWithoutParams } from '../../utils/utils';
 
 export type GalleryPropsWithContext<
@@ -311,7 +311,7 @@ const GalleryThumbnail = <
   const defaultOnPress = () => {
     // If the url is defined then only try to open the file.
     if (thumbnail.url) {
-      if (thumbnail.type === 'video' && !isVideoPlayerAvailable()) {
+      if (thumbnail.type === FileTypes.Video && !isVideoPlayerAvailable()) {
         // This condition is kinda unreachable, since we render videos as file attachment if the video
         // library is not installed. But doesn't hurt to have extra safeguard, in case of some customizations.
         openUrlSafely(thumbnail.url);
@@ -363,7 +363,7 @@ const GalleryThumbnail = <
       testID={`gallery-${invertedDirections ? 'row' : 'column'}-${colIndex}-item-${rowIndex}`}
       {...additionalTouchableProps}
     >
-      {thumbnail.type === 'video' ? (
+      {thumbnail.type === FileTypes.Video ? (
         <VideoThumbnail
           style={[
             imageBorderRadius ?? borderRadius,