+++ package/src/components/AttachmentPicker/components/AttachmentPickerItem.tsx
@@ -3,7 +3,6 @@ import React from 'react';
 import { Alert, ImageBackground, Platform, StyleSheet, Text, View } from 'react-native';
 
 import { TouchableOpacity } from '@gorhom/bottom-sheet';
-import dayjs from 'dayjs';
 import { lookup } from 'mime-types';
 
 import type { AttachmentPickerContextValue } from '../../../contexts/attachmentPickerContext/AttachmentPickerContext';
@@ -12,6 +11,7 @@ import { useViewport } from '../../../hooks/useViewport';
 import { Recorder } from '../../../icons';
 import { getLocalAssetUri } from '../../../native';
 import type { Asset, File } from '../../../types/types';
+import { getDurationLabelFromDuration } from '../../../utils/utils';
 
 type AttachmentPickerItemType = Pick<
   AttachmentPickerContextValue,
@@ -50,25 +50,15 @@ const AttachmentVideo = (props: AttachmentVideoProps) => {
 
   const { duration: videoDuration, uri } = asset;
 
-  const ONE_HOUR_IN_SECONDS = 3600;
-
-  let durationLabel = '00:00';
-
-  if (videoDuration) {
-    const isDurationLongerThanHour = videoDuration / ONE_HOUR_IN_SECONDS >= 1;
-    const formattedDurationParam = isDurationLongerThanHour ? 'HH:mm:ss' : 'mm:ss';
-    const formattedVideoDuration = dayjs
-      .duration(videoDuration, 'second')
-      .format(formattedDurationParam);
-    durationLabel = formattedVideoDuration;
-  }
+  const durationLabel = getDurationLabelFromDuration(videoDuration);
 
   const size = vw(100) / (numberOfAttachmentPickerImageColumns || 3) - 2;
 
   /* Patches video files with uri and mimetype */
   const patchVideoFile = async (files: File[]) => {
     // For the case of Expo CLI where you need to fetch the file uri from file id. Here it is only done for iOS since for android the file.uri is fine.
-    const localAssetURI = Platform.OS === 'ios' && asset.id && (await getLocalAssetUri(asset.id));
+    const localAssetURI =
+      Platform.OS === 'ios' && asset.id && getLocalAssetUri && (await getLocalAssetUri(asset.id));
     const uri = localAssetURI || asset.uri || '';
     // We need a mime-type to upload a video file.
     const mimeType = lookup(asset.name) || 'multipart/form-data';
@@ -126,7 +116,7 @@ const AttachmentVideo = (props: AttachmentVideoProps) => {
         <View style={styles.videoView}>
           <Recorder height={20} pathFill={white} width={25} />
           {videoDuration ? (
-            <Text style={[styles.durationText, durationText, { color: white }]}>
+            <Text style={[{ color: white }, styles.durationText, durationText]}>
               {durationLabel}
             </Text>
           ) : null}
@@ -162,7 +152,8 @@ const AttachmentImage = (props: AttachmentImageProps) => {
   /* Patches image files with uri */
   const patchImageFile = async (images: Asset[]) => {
     // For the case of Expo CLI where you need to fetch the file uri from file id. Here it is only done for iOS since for android the file.uri is fine.
-    const localAssetURI = Platform.OS === 'ios' && asset.id && (await getLocalAssetUri(asset.id));
+    const localAssetURI =
+      Platform.OS === 'ios' && asset.id && getLocalAssetUri && (await getLocalAssetUri(asset.id));
     const uri = localAssetURI || asset.uri || '';
     return [
       ...images,