+++ package/src/components/MessageInput/MoreOptionsButton.tsx
@@ -2,28 +2,16 @@ import React from 'react';
 import type { GestureResponderEvent } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 
-import {
-  ChannelContextValue,
-  useChannelContext,
-} from '../../contexts/channelContext/ChannelContext';
 import { useTheme } from '../../contexts/themeContext/ThemeContext';
 import { CircleRight } from '../../icons/CircleRight';
 
-import type { DefaultStreamChatGenerics } from '../../types/types';
-
-type MoreOptionsButtonPropsWithContext<
-  StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
-> = Pick<ChannelContextValue<StreamChatGenerics>, 'disabled'> & {
+export type MoreOptionsButtonProps = {
   /** Function that opens attachment options bottom sheet */
   handleOnPress?: ((event: GestureResponderEvent) => void) & (() => void);
 };
 
-const MoreOptionsButtonWithContext = <
-  StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
->(
-  props: MoreOptionsButtonPropsWithContext<StreamChatGenerics>,
-) => {
-  const { disabled, handleOnPress } = props;
+export const MoreOptionsButton = (props: MoreOptionsButtonProps) => {
+  const { handleOnPress } = props;
 
   const {
     theme: {
@@ -34,7 +22,6 @@ const MoreOptionsButtonWithContext = <
 
   return (
     <TouchableOpacity
-      disabled={disabled}
       hitSlop={{ bottom: 15, left: 15, right: 15, top: 15 }}
       onPress={handleOnPress}
       style={[moreOptionsButton]}
@@ -43,43 +30,6 @@ const MoreOptionsButtonWithContext = <
       <CircleRight pathFill={accent_blue} />
     </TouchableOpacity>
   );
-};
-
-const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics>(
-  prevProps: MoreOptionsButtonPropsWithContext<StreamChatGenerics>,
-  nextProps: MoreOptionsButtonPropsWithContext<StreamChatGenerics>,
-) => {
-  const { disabled: prevDisabled, handleOnPress: prevHandleOnPress } = prevProps;
-  const { disabled: nextDisabled, handleOnPress: nextHandleOnPress } = nextProps;
-  const disabledEqual = prevDisabled === nextDisabled;
-  if (!disabledEqual) return false;
-
-  const handleOnPressEqual = prevHandleOnPress === nextHandleOnPress;
-  if (!handleOnPressEqual) return false;
-
-  return true;
-};
-
-const MemoizedMoreOptionsButton = React.memo(
-  MoreOptionsButtonWithContext,
-  areEqual,
-) as typeof MoreOptionsButtonWithContext;
-
-export type MoreOptionsButtonProps<
-  StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
-> = Partial<MoreOptionsButtonPropsWithContext<StreamChatGenerics>>;
-
-/**
- * UI Component for more options button in MessageInput component.
- */
-export const MoreOptionsButton = <
-  StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
->(
-  props: MoreOptionsButtonProps<StreamChatGenerics>,
-) => {
-  const { disabled = false } = useChannelContext<StreamChatGenerics>();
-
-  return <MemoizedMoreOptionsButton {...{ disabled }} {...props} />;
 };
 
 MoreOptionsButton.displayName = 'MoreOptionsButton{messageInput}';