+++ package/src/components/MessageInput/components/InputEditingStateHeader.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 
-import type { ChannelContextValue } from '../../../contexts/channelContext/ChannelContext';
 import {
   MessageInputContextValue,
   useMessageInputContext,
@@ -25,19 +24,23 @@ const styles = StyleSheet.create({
   },
 });
 
-export type InputEditingStateHeaderPropsWithContext<
+export type InputEditingStateHeaderProps<
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
-> = Pick<MessageInputContextValue<StreamChatGenerics>, 'clearEditingState' | 'resetInput'> &
-  Pick<ChannelContextValue<StreamChatGenerics>, 'disabled'>;
+> = Partial<Pick<MessageInputContextValue<StreamChatGenerics>, 'clearEditingState' | 'resetInput'>>;
 
-export const InputEditingStateHeaderWithContext = <
+export const InputEditingStateHeader = <
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
 >({
-  clearEditingState,
-  disabled,
-  resetInput,
-}: InputEditingStateHeaderPropsWithContext<StreamChatGenerics>) => {
+  clearEditingState: propClearEditingState,
+  resetInput: propResetInput,
+}: InputEditingStateHeaderProps<StreamChatGenerics>) => {
   const { t } = useTranslationContext();
+  const { clearEditingState: contextClearEditingState, resetInput: contextResetInput } =
+    useMessageInputContext<StreamChatGenerics>();
+
+  const clearEditingState = propClearEditingState || contextClearEditingState;
+  const resetInput = propResetInput || contextResetInput;
+
   const {
     theme: {
       colors: { black, grey, grey_gainsboro },
@@ -54,10 +57,13 @@ export const InputEditingStateHeaderWithContext = <
         {t<string>('Editing Message')}
       </Text>
       <TouchableOpacity
-        disabled={disabled}
         onPress={() => {
-          resetInput();
-          clearEditingState();
+          if (resetInput) {
+            resetInput();
+          }
+          if (clearEditingState) {
+            clearEditingState();
+          }
         }}
         testID='close-button'
       >
@@ -65,38 +71,6 @@ export const InputEditingStateHeaderWithContext = <
       </TouchableOpacity>
     </View>
   );
-};
-
-const areEqual = <StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics>(
-  prevProps: InputEditingStateHeaderPropsWithContext<StreamChatGenerics>,
-  nextProps: InputEditingStateHeaderPropsWithContext<StreamChatGenerics>,
-) => {
-  const { disabled: prevDisabled } = prevProps;
-  const { disabled: nextDisabled } = nextProps;
-
-  const disabledEqual = prevDisabled === nextDisabled;
-  if (!disabledEqual) return false;
-
-  return true;
-};
-
-const MemoizedInputEditingStateHeader = React.memo(
-  InputEditingStateHeaderWithContext,
-  areEqual,
-) as typeof InputEditingStateHeaderWithContext;
-
-export type InputEditingStateHeaderProps<
-  StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
-> = Partial<InputEditingStateHeaderPropsWithContext<StreamChatGenerics>>;
-
-export const InputEditingStateHeader = <
-  StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics,
->(
-  props: InputEditingStateHeaderProps<StreamChatGenerics>,
-) => {
-  const { clearEditingState, resetInput } = useMessageInputContext<StreamChatGenerics>();
-
-  return <MemoizedInputEditingStateHeader {...{ clearEditingState, resetInput }} {...props} />;
 };
 
 InputEditingStateHeader.displayName = 'EditingStateHeader{messageInput}';