+++ package/src/store/apis/upsertChannelData.ts
@@ -1,8 +1,8 @@
 import type { ChannelResponse } from 'stream-chat';
 
 import { mapChannelDataToStorable } from '../mappers/mapChannelDataToStorable';
-import { SqliteClient } from '../SqliteClient';
 import { createUpsertQuery } from '../sqlite-utils/createUpsertQuery';
+import { SqliteClient } from '../SqliteClient';
 
 export const upsertChannelData = ({
   channel,